app/requirements.txt
fastapi==0.81.0
orjson==3.7.7
aioredis==2.0
elasticsearch[async]==7.9.1
pydantic==1.9.0
uvicorn==0.12.2
uvloop==0.16.0
fastapi-cache2==0.1.6
redis